requirements.txt
cycler==0.10.0
ffmpeg-python==0.2.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.1
six==1.15.0
